refactor: 테스트 리소스 제거 및 NaverUtils 생성
