--- src/main/java/com/osanvalley/moamail/domain/mail/util/SmtpMailConnector.java
@@ -1,11 +1,0 @@
-package com.osanvalley.moamail.domain.mail.util;
-
-import com.osanvalley.moamail.domain.mail.naver.dto.PostNaverMailSendRequestDto;
-
-import javax.mail.MessagingException;
-import javax.mail.Session;
-
-public interface SmtpMailConnector {
-    public Session getSession();
-    public void sendMessage(Session session, PostNaverMailSendRequestDto reqDto) throws MessagingException;
-}
