+++ src/main/java/com/osanvalley/moamail/global/imap/ImapMailConnector.java
@@ -1,4 +1,4 @@
-package com.osanvalley.moamail.domain.mail.util;
+package com.osanvalley.moamail.global.imap;
 
 import javax.mail.*;
