+++ src/main/java/com/osanvalley/moamail/global/imap/SmtpMailConnector.java
@@ -1,0 +1,11 @@
+//package com.osanvalley.moamail.global.imap;
+//
+//import com.osanvalley.moamail.domain.mail.naver.dto.PostNaverMailSendRequestDto;
+//
+//import javax.mail.MessagingException;
+//import javax.mail.Session;
+//
+//public interface SmtpMailConnector {
+//    public Session getSession();
+//    public void sendMessage(Session session, PostNaverMailSendRequestDto reqDto) throws MessagingException;
+//}
